refactor: split crypto-utils
