--- src/crypto-util/index.ts
@@ -1,1 +1,0 @@
-export * from "./rc4";
