+++ src/crypto/rc4.spec.ts
