+++ src/crypto/rc4.ts
