+++ src/crypto/types.ts
@@ -1,0 +1,4 @@
+export interface Cipher {
+    algorithm: string;
+    chaining: string;
+}
